Add initial reusable static library API
- add CMake static library target
- introduce opaque C context API for creation and rendering
- expose safe RGBA8 frame-size and readback functions
- handle WebGPU row alignment and readback resource cleanup
- add minimal consumer example covering rendering and error paths
- currently the minimal example requires display server
- plan to create tests and separate headless and windowed minimal examples

+++ include/core.h
@@ -45,7 +45,7 @@ typedef struct MouseState {
 
 WGPUBuffer create_buffer(
     WGPUDevice device,
-    const u32 size,
+    const u64 size,
     const WGPUBufferUsage usage,
     const char* label
 );
@@ -64,7 +64,7 @@ ReturnStatus load_shader(const char* path, String* buffer);
  */
 WGPUBuffer create_buffer(
     WGPUDevice device,
-    const u32 size,
+    const u64 size,
     const WGPUBufferUsage usage,
     const char* label
 ) {
